+++ NEWS.md
@@ -1,5 +1,8 @@
 # tidyselect (development version)
 
+* `matches()` recognises and correctly uses stringr pattern objects
+  (`stringr::regexp()`, `stringr::fixed()`, etc) (#238). 
+
 * `eval_select(allow_rename = FALSE)` no longer fails with empty
   selections (#221, @eutwt).